Randomizer can be used with DB

+++ spa-backend/src/main/java/com/bigshots/spabackend/repo/JokeRepo.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import com.bigshots.spabackend.model.Joke;
 
 @Repository
-public interface JokeRepo extends JpaRepository<Joke, Long>{
+public interface JokeRepo extends JpaRepository<Joke, Integer>{
 
 }
